requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
blinker==1.8.2
    # via flask
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
flask-sqlalchemy==3.1.1
    # via -r requirements.in
idna==3.7
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.2
    # via -r requirements.in
sqlalchemy==2.0.30
    # via flask-sqlalchemy
typing-extensions==4.11.0
    # via sqlalchemy
urllib3==2.2.1
    # via requests
werkzeug==3.0.3
    # via flask
